requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.7
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.9
djangorestframework==3.12.4
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
urllib3==1.26.7
